# lib/fs: More efficient casefs cache