--- a/lib/fs/debug_symlink_unix.go
+++ b/lib/fs/debug_symlink_unix.go
@@ -29,19 +29,3 @@ func DebugSymlinkForTestsOnly(oldFs, newFs Filesystem, oldname, newname string) 
 	}
 	return nil
 }
-
-// unwrapFilesystem removes "wrapping" filesystems to expose the underlying filesystem.
-func unwrapFilesystem(fs Filesystem) Filesystem {
-	for {
-		switch sfs := fs.(type) {
-		case *logFilesystem:
-			fs = sfs.Filesystem
-		case *walkFilesystem:
-			fs = sfs.Filesystem
-		case *MtimeFS:
-			fs = sfs.Filesystem
-		default:
-			return sfs
-		}
-	}
-}