@imsodin 2020-09-09
Should look at the failing check before approving: `unwrapFilesystem` is currently only used/defined behind `// +build !windows`